requirements.txt
langchain>=0.2.15
langgraph>=0.2.33
langchain-openai>=0.1.22
python-dotenv>=1.0.1
pandas>=2.2.2
